+++ onnxscript/optimizer/_legacy/constant_folding.py
@@ -10,10 +10,9 @@ import onnx
 import onnx.reference.ops
 
 import onnxscript._legacy_ir as ir
-from onnxscript._legacy_ir import visitor
-import onnxscript.optimizer as optimizer
-from onnxscript.optimizer._legacy import evaluator
 import onnxscript.optimizer._constant_folding as _constant_folding
+from onnxscript._legacy_ir import visitor
+from onnxscript.optimizer._legacy import evaluator
 from onnxscript.utils.utils import (
     is_control_flow_op,
     is_onnx_domain,
